services/time_tracking_service/requirements.txt
Django>=5.1,<5.2
djangorestframework>=3.15,<4.0
djangorestframework-simplejwt>=5.3,<6.0
django-filter>=24.0,<25.0
mysqlclient>=2.2,<3.0
celery>=5.4,<6.0
redis>=5.0,<6.0
drf-spectacular>=0.27,<1.0
gunicorn>=22.0,<23.0
python-decouple>=3.8,<4.0
requests>=2.32,<3.0
PyJWT>=2.8,<3.0
pytest>=8.0,<9.0
pytest-django>=4.8,<5.0
